+++ src/app/dashboard/components/widgets-grid/widgets-grid.component.ts
@@ -241,6 +241,7 @@ export class WidgetsGridComponent implements OnInit, OnDestroy {
     public isSmpHeader(widgetType: string): boolean {
         switch (widgetType) {
             case 'cd-critical':
+            case 'cmid-filter':
                 return true;
         }
         return false;