+++ src/app/dashboard/models/KPE/kpe-accuracy-timelines-data.model.ts
@@ -1,0 +1,5 @@
+export interface IKpeAccTimelinesDataEditPlan {
+    adjustmentComment: string,
+    dateOfCreation: string,
+    dateOfCorrection: string
+}