+++ src/app/widgets/EC/ec-widget-shared/constants/colors.const.ts
@@ -1,7 +1,7 @@
 export const lineColors: { [key: string]: string } = {
     1: 'var(--data-c5-color)',
     2: 'var(--index-plan-color)',
-    3: 'var(--index-blue1-color)',
+    3: 'var(--data-c9-color)',
     4: 'var(--data-c3-color)',
     5: 'var(--data-c7-color)',
     6: 'var(--data-c1-color)',