+++ src/app/widgets/SOU/sou-streams/sou-streams.component.ts
@@ -4,14 +4,15 @@ import {
     Inject,
     ViewChild,
     ElementRef,
-    AfterViewChecked, ChangeDetectorRef
+    AfterViewChecked, ChangeDetectorRef, ViewContainerRef, TemplateRef, AfterViewInit, OnDestroy
 } from '@angular/core';
 import { WidgetService } from '@dashboard/services/widget.service';
 import { WidgetPlatform } from '@dashboard/models/@PLATFORM/widget-platform';
-
+import { Overlay, OverlayRef } from '@angular/cdk/overlay';
+import { TemplatePortal } from '@angular/cdk/portal';
 import { TITLES_OF_TABLE } from '@widgets/SOU/sou-streams/config';
-import { TABLE_CELLS } from '@widgets/SOU/sou-streams/mock';
 import { animate, state, style, transition, trigger } from '@angular/animations';
+import { SouStreamsService } from '@dashboard/services/widgets/SOU/sou-streams.service';
 
 
 @Component({
@@ -42,22 +43,32 @@ import { animate, state, style, transition, trigger } from '@angular/animations'
         )
     ]
 })
-export class SouStreamsComponent extends WidgetPlatform implements OnInit, AfterViewChecked {
+export class SouStreamsComponent extends WidgetPlatform implements OnInit, AfterViewChecked, AfterViewInit, OnDestroy {
 
     public titlesOfTable: { name: string, bigBlock?: boolean }[] = TITLES_OF_TABLE;
-    public tableRows: {} = TABLE_CELLS;
+    public tableRows: {};
+
+    public isReservoirTrendOpen: boolean = false;
 
     public heightOfTable: string = '400px';
     public heightOfViewPort: string = '335px';
     public widthOfTable: string = '1943.2px';
     public widthOfGraphic: number = 70;
+    @ViewChild('widget') child: ElementRef;
+
+    @ViewChild('overlayCustom') dialogTemplate: TemplateRef<any>;
+    private overlayRef: OverlayRef;
+    private portal: TemplatePortal;
 
     public showWorkspace: boolean = false;
     public show: boolean = true;
 
     constructor(
         private cdr: ChangeDetectorRef,
-        protected widgetService: WidgetService,
+        public widgetService: WidgetService,
+        public souStreamsService: SouStreamsService,
+        private overlay: Overlay,
+        private viewContainerRef: ViewContainerRef,
         @Inject('widgetId') public id: string,
         @Inject('uniqId') public uniqId: string
     ) {
@@ -66,9 +77,38 @@ export class SouStreamsComponent extends WidgetPlatform implements OnInit, After
 
     ngOnInit(): void {
         super.widgetInit();
+        this.souStreamsService.getTableContent('45', '67').then((res) => {
+            this.tableRows = res;
+        });
     }
 
-    @ViewChild('widget') child: ElementRef;
+    processDate(inputData: string): string {
+        return `${inputData.slice(8, 10)}.${inputData.slice(5, 7)}.${inputData.slice(0, 4)},
+        ${inputData.slice(11, 16)}`;
+    }
+
+    processWarningColor(inputColor: string): string {
+        return (inputColor === 'blue') ? 'var(--color-table-line-blue)' :
+            (inputColor === 'yellow') ? 'var(--color-sou-orange)' : 'var(--bg-body-color)';
+    }
+
+    ngAfterViewInit(): void {
+        this.portal = new TemplatePortal(this.dialogTemplate, this.viewContainerRef);
+        this.overlayRef = this.overlay.create({
+            positionStrategy: this.overlay.position().global().centerHorizontally().centerVertically(),
+            hasBackdrop: true
+        });
+        this.overlayRef.backdropClick().subscribe(() => this.overlayRef.detach());
+    }
+
+    overlayDetach(): void {
+        this.overlayRef.detach();
+    }
+
+    ngOnDestroy(): void {
+        this.overlayRef.dispose();
+    }
+
 
     ngAfterViewChecked(): void {
         this.heightOfTable = this.child.nativeElement.clientHeight - 40 + 'px';
@@ -103,4 +143,17 @@ export class SouStreamsComponent extends WidgetPlatform implements OnInit, After
         this.heightOfViewPort = this.child.nativeElement.clientHeight - 115 - height + 'px';
         this.cdr.detectChanges();
     }
+    public openReservoirTrend(): void {
+        this.isReservoirTrendOpen = true;
+    }
+
+    public closeSmartTrend(): void {
+        this.isReservoirTrendOpen = false;
+        this.overlayDetach();
+    }
+
+    openDialog(): void {
+        this.overlayRef?.attach(this.portal);
+    }
+
 }