+++ src/app/widgets/SUUTP/suutp-complexes-table/suutp-complexes-table.interface.ts
@@ -1,0 +1,18 @@
+export interface ITableDataRow {
+    value: string;
+    isWarning?: boolean;
+}
+
+export interface ITableData {
+    cells: any[];
+}
+
+export interface ITableHeaders {
+    [key: string]: string;
+}
+
+export interface ITableMain {
+    headers: ITableHeaders[];
+    rows: any[];
+    displayedColumns: string[];
+}